feature/COR-1954-consolidate-metadata (#5007)
* feat(COR-1954): Initial copy-paste merging

* feat(COR-1954): Add JSDoc comments

* feat(COR-1954): Fix empty source showing up

* feat(COR-1954): Replace source with metadataItem

* feat(COR-1954): Turn datelastinserted date into MetadataItem

* feat(COR-1954): Consolidated metadataprops type into seperate type file

* feat(COR-1954) Split up and combine into module

* feat(COR-1954): Substitute clock on pageinfoblock with metadataitem
