+++ packages/app/src/components/kpi/bordered-kpi-section.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { Box } from '~/components/base';
 import { mediaQueries, space } from '~/style/theme';
 import { KpiTile } from '../kpi-tile';
-import { MetadataProps } from '../metadata';
+import { MetadataProps } from '~/components/metadata';
 import { TwoKpiSection } from '../two-kpi-section';
 import { KpiContent } from './components/kpi-content';
 import { BorderedKpiSectionProps } from './types';
