+++ packages/app/src/components/metadata/index.ts
@@ -1,0 +1,2 @@
+export * from './metadata';
+export * from './types';
