+++ packages/app/src/components/page-information-block/page-information-block.tsx
@@ -12,7 +12,7 @@ import { mediaQueries, radii, space } from '~/style/theme';
 import { RichContentBlock } from '~/types/cms';
 import { useScopedWarning } from '~/utils/use-scoped-warning';
 import { Header } from './components/header';
-import { Metadata, MetadataProps } from './components/metadata';
+import { Metadata, MetadataProps } from '~/components/metadata';
 import { PageLinks } from './components/page-links';
 
 interface InformationBlockProps {
@@ -75,7 +75,7 @@ export function PageInformationBlock({
 
   const MetaDataBlock = metadata ? (
     <MetadataBox>
-      <Metadata {...metadata} accessibilitySubject={title} referenceLink={referenceLink} />
+      <Metadata {...metadata} accessibilitySubject={title} referenceLink={referenceLink} isPageInformationBlock />
     </MetadataBox>
   ) : null;
